This notebook shows a clean EDA and ML pipeline
See this notebook for a clean EDA of growth measurements that are then put through binary classification to predict cancer using a Multi Layer Perceptron type of Neural Network. The primary point of this notebook is what a clean ML pipeline can provide. Several alternative ML algorithms or Neural Network designs should be added and the best fitting used in an ensemble.